# Optimize unnecessary check in VecDeque::retain