--- a/library/alloc/src/collections/vec_deque/tests.rs
+++ b/library/alloc/src/collections/vec_deque/tests.rs
@@ -87,8 +87,6 @@ fn bench_pop_front_100(b: &mut test::Bencher) {
     })
 }
 
-
-
 #[test]
 fn test_swap_front_back_remove() {
     fn test(back: bool) {