@TennyZhuang 2021-08-16
I ran the benchmark on my Mac

```
sysctl -n machdep.cpu.brand_string
Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
```

Before optimization:

```
test collections::vec_deque::tests::bench_retain_half_10000  ... bench:     397,713 ns/iter (+/- 19,098)
test collections::vec_deque::tests::bench_retain_odd_10000   ... bench:     382,615 ns/iter (+/- 17,406)
test collections::vec_deque::tests::bench_retain_whole_10000 ... bench:     386,554 ns/iter (+/- 11,450)
```

After optimization:

```
test collections::vec_deque::tests::bench_retain_half_10000  ... bench:     318,491 ns/iter (+/- 34,193)
test collections::vec_deque::tests::bench_retain_odd_10000   ... bench:     354,253 ns/iter (+/- 21,848)
test collections::vec_deque::tests::bench_retain_whole_10000 ... bench:     288,971 ns/iter (+/- 16,016)
```